# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "List members",
                "description": "List all resgistered members.",
                "operationId": "0a7045fbb9a3c713742cb2ed937a0105",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "member_id": 1,
                                        "member_name": "Cicero Augusto",
                                        "member_email": "ciceroam@gmail.com"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/members/email/{email}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Search member by email",
                "description": "Gets the data of a member through an email.",
                "operationId": "872ed260f5b548ca771eab14113f9647",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "member_id": 1,
                                        "member_name": "Cicero Augusto",
                                        "member_email": "ciceroam@gmail.com"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/members/noapplication": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "List members with no applications",
                "description": "List all resgistered members with no job applications.",
                "operationId": "94e7cd5706e449bda36e5e76a6ba2a70",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "member_id": 1,
                                        "member_name": "Cicero Augusto",
                                        "member_email": "ciceroam@gmail.com"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/resumes": {
            "get": {
                "tags": [
                    "Resumes"
                ],
                "summary": "List resumes",
                "description": "List all resgistered resumes.",
                "operationId": "e8504de846fba76c6d19df4f0d767320",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "resume_id": 1,
                                        "resume_candidate": "Cicero Augusto",
                                        "resume_description": "System Analyst with experience"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/resumes/name/{name}": {
            "get": {
                "tags": [
                    "Resumes"
                ],
                "summary": "Search resume by candidate name",
                "description": "Gets the data of a resume through a candidate name.",
                "operationId": "702b2d0c36ff9705b59c55c92ebae156",
                "parameters": [
                    {
                        "name": "Name",
                        "in": "path",
                        "description": "Resume Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "resume_id": 1,
                                        "resume_candidate": "Cicero Augusto",
                                        "resume_description": "System Analyst with experience"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}